Create `organizations` module and put organization-specific dashboard in `organizations/dashboards_controller`
