+++ app/controllers/organizations/base_controller.rb
@@ -1,0 +1,5 @@
+module Organizations
+  class BaseController < ApplicationController
+    before_action :authenticate_user!
+  end
+end
