+++ app/views/aid_applications/verifications/edit.html.erb
@@ -57,7 +57,7 @@
         <div class="toolbar">
           <div class="toolbar__left"></div>
           <div class="toolbar__right">
-            <%= link_to t('general.submit'), organization_path(current_organization), class: 'toolbar__item button button--primary' %>
+            <%= link_to t('general.submit'), organization_dashboard_path(current_organization), class: 'toolbar__item button button--primary' %>
           </div>
         </div>
       </div>
