+++ app/views/organizations/dashboards/show.html.erb
@@ -4,7 +4,7 @@
 <div class="grid">
   <div class="grid__item width-one-whole">
     <%= button_to organization_aid_applications_path(current_organization), method: :post, class: 'button button--primary' do %>
-      <span class="icon icon-add" aria-hidden="true"></span> <%= t('aid_applications.index.add_app') %>
+        <span class="icon icon-add" aria-hidden="true"></span> <%= t('aid_applications.index.add_app') %>
     <% end %>
   </div>
 </div>
