+++ app/views/shared/_header.erb
@@ -9,7 +9,7 @@
       <div class="toolbar__right">
         <% if current_user.present? %>
           <% if current_user.organization.present? %>
-            <%= link_to 'Dashboard', organization_path(current_user.organization), class: 'toolbar__item text--small' %>
+            <%= link_to 'Dashboard', organization_dashboard_path(current_user.organization), class: 'toolbar__item text--small' %>
           <% else %>
             <%= link_to 'Admin', organizations_path, class: 'toolbar__item text--small' %>
           <% end %>
